examples/python/facilitator/requirements.txt
x402>=0.1.0
python-dotenv>=1.0.0
tronpy>=0.4.0
